# [TINKERPOP-3004] improve GremlinValueComparator performance