--- a/gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/util/GremlinValueComparator.java
+++ b/gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/util/GremlinValueComparator.java
@@ -129,6 +129,9 @@ public abstract class GremlinValueComparator implements Comparator<Object> {
             if (containersOfDifferentSize(f, s))
                 return false;
 
+            if (f != null && s != null && f.hashCode() != s.hashCode())
+                return false;
+
             // For Compare, NaN always produces ERROR
             if (eitherAreNaN(f, s))
                 return false;
@@ -138,9 +141,9 @@ public abstract class GremlinValueComparator implements Comparator<Object> {
                 return false;
 
             try {
-              // comparable(f, s) assures that type(f) == type(s)
-              final Type type = Type.type(f);
-              return comparator(type).compare(f, s) == 0;
+                // comparable(f, s) assures that type(f) == type(s)
+                final Type type = Type.type(f);
+                return comparator(type).compare(f, s) == 0;
             } catch (GremlinTypeErrorException ex) {
                 /**
                  * By routing through the compare(f, s) path we expose ourselves to type errors, which should be
@@ -151,7 +154,7 @@ public abstract class GremlinValueComparator implements Comparator<Object> {
                  *
                  * Can also happen for elements nested inside of collections.
                  */
-               return false;
+                return false;
             }
         }