--- a/gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/util/GremlinValueComparator.java
+++ b/gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/util/GremlinValueComparator.java
@@ -129,7 +129,9 @@ public abstract class GremlinValueComparator implements Comparator<Object> {
             if (containersOfDifferentSize(f, s))
                 return false;
 
-            if (f != null && s != null && f.hashCode() != s.hashCode())
+            // numbers and collections with different hashcode can be equal
+            if (f != null && s != null && f.hashCode() != s.hashCode() && f.hashCode() != s.hashCode()
+                    && !(f instanceof Number) && !(f instanceof Collection) && !(f instanceof Map))
                 return false;
 
             // For Compare, NaN always produces ERROR