@Cole-Greer 2023-10-17
Thanks Valentyn, it's always nice to see big gains in performance. Hopefully we can get back closer to 3.5 performance at some point but this is a big step forward.

VOTE +1